Add BaseHtmlLayout and move htmlContent there. Update other classes and tests accordingly.
