+++ universal-application-tool-0.0.1/app/views/LoginForm.java
@@ -4,13 +4,19 @@ import static j2html.TagCreator.body;
 import static j2html.TagCreator.form;
 import static j2html.TagCreator.h1;
 
+import akka.parboiled2.RuleTrace;
 import play.mvc.Http;
 import play.twirl.api.Content;
 
 public class LoginForm extends BaseHtmlView {
+  private final BaseHtmlLayout layout;
+
+  public LoginForm(BaseHtmlLayout layout) {
+    this.layout = layout;
+  }
 
   public Content render(Http.Request request) {
-    return htmlContent(
+    return layout.htmlContent(
         body(
             h1("Log In"),
             form(
